requirements.txt
python-dotenv~=1.0.1
geoip2~=4.8.0